Extend NettyHttpServerTransport instead of copying it
I think back when we initially created the blob stuff we had to copy the
`NettyHttpServerTransport` because we depended on a shaded ES jar and
couldn't use certain netty classes.

This is no longer the case, so it's possible to just extend the
NettyHttpServerTransport.
